+++ tests/testthat/test-select-helpers.R
@@ -1,4 +1,16 @@
 context("select-helpers")
+
+test_that("failed match removes all columns", {
+  set_current_vars(c("x", "y"))
+  on.exit(reset_current_vars())
+
+  expect_equal(starts_with("z"), -(1:2))
+  expect_equal(ends_with("z"), -(1:2))
+  expect_equal(contains("z"), -(1:2))
+  expect_equal(matches("z"), -(1:2))
+  expect_equal(num_range("z", 1:3), -(1:2))
+})
+
 # one_of ------------------------------------------------------------------
 
 test_that("one_of gives useful errors", {